chore(deps): Update .NET SDK version and clean up Dockerfile
This commit updates the .NET SDK version to 9.0.304 in global.json.

Additionally, it removes the redundant cache-priming steps from the Dockerfile. The temporary creation and restoration of console and class library projects are no longer necessary, as the dotnet restore step is handled directly within the CI/CD pipeline. This change reduces unnecessary image bloat and streamlines the Docker build process.